chore: process api coverage in parallel
