+++ .circleci/config.yml
@@ -107,6 +107,7 @@ jobs:
             TESTFILES=$(circleci tests glob "test/**/*.spec.ts" | circleci tests split --split-by=timings)
             yarn test $TESTFILES --coverageDirectory=reports --reporters=jest-junit --reporters=default
             mv reports/coverage-final.json reports/coverage-${CIRCLE_NODE_INDEX:-0}.json
+            mv api-coverage/api-coverage.json api-coverage/api-coverage-${CIRCLE_NODE_INDEX:-0}.json
           environment:
             CI: 'true'
             JEST_JUNIT_OUTPUT_DIR: './test-result'
@@ -119,6 +120,7 @@ jobs:
           root: *workspace_root
           paths:
             - content/reports
+            - content/api-coverage
 
   report-content-coverage:
     <<: *base_env
@@ -136,6 +138,15 @@ jobs:
           path: content/coverage
       - store_artifacts:
           path: content/coverage
+  report-content-api-coverage:
+    <<: *base_env
+    steps:
+      - <<: *attach_root
+      - run:
+          name: Report API coverage
+          command: |
+            cd content
+            yarn report-api-coverage
   build-lambdas:
     <<: *base_env
     steps:
@@ -248,6 +259,10 @@ workflows:
           <<: *all_branches_and_tags
           requires:
             - build-content
+      - report-content-api-coverage:
+          <<: *all_branches_and_tags
+          requires:
+            - build-content
       - build-lambdas:
           <<: *all_branches_and_tags
           requires:
@@ -266,6 +281,7 @@ workflows:
             - build-lighthouse
             - build-content
             - report-content-coverage
+            - report-content-api-coverage
             - build-lambdas
 
       # NOT (main | semver) enables manual approval to release untagged docker images
