+++ .gitignore
@@ -1,6 +1,7 @@
 
 .env
 coverage
+api-coverage
 dist
 **/node_modules
 yarn-error.log
