+++ content/reportApiCoverage.js
@@ -1,0 +1,3 @@
+const printApiCoverage = require('./printApiCoverage')
+
+printApiCoverage()
